completely changed the data layer to sql compact (it should have been this way since the first release); minor other changes; release comes maybe after some further tests

+++ TS3-Bot.Core/Manager/Features/AwayManager.cs
@@ -123,8 +123,8 @@
                 if (!Repository.Channel.GetClientSticky(client.ClientDatabaseId).HasValue)
                 {
                     var awayClient = Repository.Client.GetLastChannelByClientId(client.ClientDatabaseId);
-                    var channel = Repository.Channel.GetChannelListInfo(awayClient.LastChannelId);
-                    QueryRunner.MoveClient(client.ClientId, awayClient.LastChannelId);
+                    var channel = Repository.Channel.GetChannelListInfo((uint)awayClient.LastChannelId);
+                    QueryRunner.MoveClient(client.ClientId, (uint)awayClient.LastChannelId);
 
                     Log(Repository.Settings.Away,
                         string.Format("Client '{0}'(id:{1}) successfully moved back from Away Channel to '{2}'(id:{3}).",
@@ -138,7 +138,7 @@
                             ClientDatabaseId = client.ClientDatabaseId,
                             ClientNickname = client.Nickname,
                             ClientAwayTime = BasicHelper.GetTimespanString(awayTimespan),
-                            ChannelId = awayClient.LastChannelId,
+                            ChannelId = (uint)awayClient.LastChannelId,
                             ChannelName = channel.Name
                         };
                         QueryRunner.SendTextMessage(MessageTarget.Server, Repository.Connection.CredentialEntity.Self.VirtualServerId, Repository.Settings.Away.TextMessage.ToMessage(messageContext));
@@ -166,8 +166,8 @@
                 if (!Repository.Channel.GetClientSticky(client.ClientDatabaseId).HasValue)
                 {
                     var idleClient = Repository.Client.GetLastChannelByClientId(client.ClientDatabaseId);
-                    var channel = Repository.Channel.GetChannelListInfo(idleClient.LastChannelId);
-                    QueryRunner.MoveClient(client.ClientId, idleClient.LastChannelId);
+                    var channel = Repository.Channel.GetChannelListInfo((uint)idleClient.LastChannelId);
+                    QueryRunner.MoveClient(client.ClientId, (uint)idleClient.LastChannelId);
 
                     Log(Repository.Settings.Idle,
                         string.Format("Client '{0}'(id:{1}) successfully moved back from Idle Channel to '{2}'(id:{3}).",
@@ -181,7 +181,7 @@
                             ClientDatabaseId = client.ClientDatabaseId,
                             ClientNickname = client.Nickname,
                             ClientAwayTime = BasicHelper.GetTimespanString(idleTimespan),
-                            ChannelId = idleClient.LastChannelId,
+                            ChannelId = (uint)idleClient.LastChannelId,
                             ChannelName = channel.Name
                         };
                         QueryRunner.SendTextMessage(MessageTarget.Server, Repository.Connection.CredentialEntity.Self.VirtualServerId, Repository.Settings.Idle.TextMessage.ToMessage(messageContext));
